pyproject.toml
[project]
name = "multi-agent-research-lab"
version = "0.1.0"
description = ""
authors = [
    {name = "JavierRodriguezFontiveros",email = "javierrodriguezfontiveros@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "python-dotenv>=1.0.1",
    "langchain>=0.3.0",          
    "langchain-community>=0.3.0", 
    "langgraph>=0.2.0",
    "langchain-ollama>=0.2.0",    
    "langfuse (>=3.14.5,<4.0.0)",
    "transformers>=4.40",
    "accelerate>=0.27",
    "torch>=2.2",
    "langchain-core>=0.3.0",      
    "langchain-chroma>=0.1.4",    
    "chromadb>=0.5.0",            
    "pypdf>=5.0.0",              
    "langchain-huggingface>=0.1.0",              
    "sentence-transformers (>=5.2.3,<6.0.0)",              
    "duckduckgo-search (>=8.1.1,<9.0.0)",              
    "langfuse-langchain (>=2.60.10.1,<3.0.0.0)",              
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"